% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_modelling.R
\name{add_int}
\alias{add_int}
\title{Add interaction columns to data frame}
\usage{
add_int(.data, .formula)
}
\arguments{
\item{.data}{dataframe}

\item{.formula}{right hand sided formula}
}
\value{
dataframe
}
\description{
Add interaction columns to data frame
}
\author{
Søren Højsgaard
}
