% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BY-funBy.R
\name{by-fun}
\alias{by-fun}
\alias{formulaFunBy}
\alias{xyFunBy}
\title{A template function for creating groupwise functions}
\usage{
formulaFunBy(formula, group, data, FUN, class = NULL, ...)

xyFunBy(xy, group, data, FUN, class = NULL, ...)
}
\arguments{
\item{formula}{A formula of the form \code{y ~ x} (which must be variable
names in \code{data}).}

\item{group}{A right hand sided formula or a character vector defining the
grouping of data}

\item{data}{A data frame}

\item{FUN}{The function to be applied}

\item{class}{The class to give the result of the returned value of the
created function.}

\item{xy}{A character vector with one or two elements (which must be variable
names in \code{data}).}

\item{\dots}{Further arguments passed on to \code{FUN}}
}
\value{
A function
}
\description{
A template function for creating groupwise functions
}
\note{
This function is a recent addition and has not been thoroughly tested.
    Please report bugs.
}
\examples{


## Example: Create a function for creating groupwise t-tests

mydata <- data.frame(y=rnorm(32), x=rnorm(32),
g1=factor(rep(c(1,2),each=16)), g2=factor(rep(c(1,2), each=8)),
g3=factor(rep(c(1,2),each=4))) 


t.testBy1 <- function(formula, group, data, ...){
  formulaFunBy(formula, group, data, FUN=t.test, class="t.testBy1", ...)
}

t.testBy2 <- function(formula, group, data, ...){
  xyFunBy(formula, group, data, FUN=t.test, class="t.testBy1", ...)
}


t.testBy1(y~g1, ~g2+g3, data=mydata)
t.testBy2(y~x, ~g2+g3, data=mydata)


}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\seealso{
\code{\link{parseGroupFormula}}
}
\keyword{utilities}

