% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BY-subsetBy.R
\name{by-subset}
\alias{by-subset}
\alias{subsetBy}
\alias{subset_by}
\title{Finds subsets of a dataframe which is split by variables in a formula.}
\usage{
subsetBy(formula, subset, data = parent.frame(), select, drop = FALSE,
  join = TRUE, ...)

subset_by(data, formula, subset, select, drop = FALSE, join = TRUE, ...)
}
\arguments{
\item{formula}{A formula to split by.}

\item{subset}{logical expression indicating elements or rows to keep: missing
values are taken as false.}

\item{data}{A data frame.}

\item{select}{expression, indicating columns to select from a data frame.}

\item{drop}{passed on to [ indexing operator.}

\item{join}{If FALSE the result is a list of data frames (as defined by
'formula'); if TRUE one data frame is returned.}

\item{...}{further arguments to be passed to or from other methods.}
}
\value{
A data frame.
}
\description{
A data frame is split by a formula into groups. Then subsets are found
within each group, and the result is collected into a data frame.
}
\examples{

data(dietox)
subsetBy(~Evit, Weight < mean(Weight), data=dietox)

}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\seealso{
See Also \code{\link{splitBy}}
}
\keyword{utilities}

