% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_doby.R
\docType{data}
\name{child_growth}
\alias{child_growth}
\title{Berkeley Growth Study data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2883 rows and 4 columns.
}
\usage{
child_growth
}
\description{
dataframe with heights of 39 boys and 54 girls from
age 1 to 18 and the ages at which they were collected.

Format:

gender: Gender of child
age: Age at time of data recordning
subject: Idenfication for each child
height: Height of child
}
\details{
Notice that the ages are not equally spaced. Data are
taken from the fda package (growth) but put in long format
here.
}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Ramsay, James O., Hooker, Giles, and Graves, Spencer (2009),
_Functional data analysis with R and Matlab_, Springer, New York.

Ramsay, James O., and Silverman, Bernard W. (2005), _Functional
Data Analysis, 2nd ed._, Springer, New York.

Ramsay, James O., and Silverman, Bernard W. (2002), _Applied
Functional Data Analysis_, Springer, New York.

Tuddenham, R. D., and Snyder, M. M. (1954) "Physical growth of
California boys and girls from birth to age 18", _University of
California Publications in Child Development_, 1, 183-364.
}\if{html}{\out{</div>}}
}
\keyword{datasets}
