% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_op.R
\name{chr_to_matrix}
\alias{chr_to_matrix}
\title{Character vector to matrix}
\usage{
chr_to_matrix(x, value = 0)
}
\arguments{
\item{x}{character vector}

\item{value}{value in matrix}
}
\description{
Character vector to matrix
}
\details{
creates square matrix with \code{x} as row and column names and \code{val} as values
}
\examples{

d1 <- letters[1:3]
chr_to_matrix(d1, 3:5)

}
