% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_doby.R
\docType{data}
\name{crimeRate}
\alias{crimeRate}
\title{crimeRate}
\format{
This data frame contains:
\describe{
\item{state:}{State of the USA}
\item{murder:}{crime of murder}
\item{rape:}{}
\item{robbery:}{}
\item{assault:}{}
\item{burglary:}{residential theft}
\item{larceny:}{unlawful taking of personal property (pocket picking)}
\item{autotheft:}{}
}
}
\usage{
crimeRate
}
\description{
Crime rates per 100,000 inhabitants in states of the USA for different crime
types in 1977.
}
\examples{
	
data(crimeRate)

}
\keyword{datasets}
