% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_doby.R
\docType{data}
\name{cropyield}
\alias{cropyield}
\title{Yield from Danish agricultural production of grain and root crop.}
\format{
A dataframe with 97 rows and 7 columns.
\describe{
\item{\code{year}}{From 1901 to 1997.}
\item{\code{precip}}{Milimeter precipitation.}
\item{\code{yield}}{Million feed units (see details).}
\item{\code{area}}{Area in 1000 ha for grains and root crop.}
\item{\code{fertil}}{1000 tons fertilizer.}
\item{\code{avgtmp1}}{Average temperature April-June (3 months).}
\item{\code{avgtmp2}}{Average temperature July-Octobre (4 months). }
}
}
\usage{
cropyield
}
\description{
Yield from Danish agricultural production of grain and root crop.
}
\details{
A feed unit is the amount of energy in a kg of barley.
}
\references{
Danmarks statistik (Statistics Denmark).
}
\keyword{datasets}
