% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-potatoes.R
\docType{data}
\name{data-potatoes}
\alias{data-potatoes}
\alias{potatoes}
\title{Weight and size of 20 potatoes}
\format{A data frame with 20 observations on the following 3 variables.
  \describe{
    \item{\code{weight}}{a numeric vector}
    \item{\code{length}}{a numeric vector}
    \item{\code{width}}{a numeric vector}
  }}
\source{
My own garden; autumn 2015.
}
\usage{
potatoes
}
\description{
Weight and size of 20 potatoes. Weight in grams; size in milimeter. There
are two sizes: \code{length} is the longest length and \code{width} is the
shortest length across a potato. #'
}
\examples{

data(potatoes)
plot(potatoes) 

}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\keyword{datasets}
