% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_doby.R
\docType{data}
\name{data_fev}
\alias{data_fev}
\alias{fev}
\title{Forced expiratory volume in children}
\format{
A data frame with 654 observations on the following 5 variables.
\describe{
\item{\code{Age}}{Age in years.}
\item{\code{FEV}}{Forced expiratory volume in liters per second.}
\item{\code{Ht}}{Height in centimeters.}
\item{\code{Gender}}{Gender.}
\item{\code{Smoke}}{Smoking status.}
}
}
\usage{
fev
}
\description{
Dataset to examine if respiratory function in children
was influenced by smoking.
}
\examples{

data(fev)
summary(fev)

}
\references{
I. Tager and S. Weiss and B. Rosner and F. Speizer (1979). Effect
of Parental Cigarette Smoking on the Pulmonary Function of
Children.  American Journal of Epidemiology. 110:15-26
}
\keyword{datasets}
