% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_doby.R
\docType{data}
\name{data_income}
\alias{data_income}
\alias{income}
\title{Income data, years of educations and ethnicity}
\format{
This data frame contains:
\describe{
\item{inc:}{Income: Yearly income (thousands of dollars).}
\item{educ:}{Education: Number of years of education (12=high school graduate, 16=college graduate).}
\item{race:}{Racial-Ethnic group: "b" (black), "h" (hispanic) and "w" (white).}
}
}
\usage{
income
}
\description{
Data on income, years of educations and ethnicity for a samle of adult Americans aged over 25. The year of sampling is not avalable in the source.
}
\details{
Variable names are as in the reference.
}
\references{
Agresti, A. (2024) Statistical Methods for the Social
Sciences, Global Edition (6th edition).  ISBN-13: 9781292449197. Table 13.1
}
\concept{dataset}
\keyword{datasets}
