% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_doby.R
\docType{data}
\name{data_math_teachers}
\alias{data_math_teachers}
\alias{math_teachers}
\title{Height of math teachers}
\format{
:
\describe{
\item{height}{Height in centimeters}
\item{sex}{Male or female}
}
}
\usage{
math_teachers
}
\description{
Height of a sample of math teachers in Danish high
schools collected at a continued education day at Mariager Fjord Gymnasium in 2019.
}
\examples{

aggregate(height ~  sex, data=math_teachers, FUN=mean)
aggregate(height ~  sex, data=math_teachers, FUN=function(x) {c(mean=mean(x), sd=sd(x))})

}
\concept{dataset}
\keyword{datasets}
