% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_doby.R
\docType{data}
\name{data_nir_milk}
\alias{data_nir_milk}
\alias{nir_milk}
\alias{NIRmilk}
\title{Near infra red light (NIR) measurements in milk}
\format{
Data comes in two formats:

nir_milk: A list with two components
\itemize{
\item x Datafrane with infra red light
amount at different wavelengths (column names are the wavelengths;
just remove the leading X).
\item y Datafrane with response variables
fat, protein, lactose and dm (drymatter)
}

NIRmilk: This data frame contains 17 rows and 158 columns.
\itemize{
\item The first column is the sample number.
\item The columns named in the form \code{Xw} contains
the transmittance (fraction of electromagnetic power)
transmittance through the sample at wavelength \code{w}.
\item The response variables are fat, protein, lactose and dm (dry
matter).
}

An object of class \code{data.frame} with 17 rows and 158 columns.
}
\usage{
nir_milk

NIRmilk
}
\description{
Near infra red light (NIR) measurements are made at
152 wavelengths on 17 milk samples.  While milk runs through a
glass tube, infra red light is sent through the tube and the
amount of light passing though the tube is measured at
different wavelengths.  Each milk sample was additionally
analysed for fat, lactose, protein and dry matter.
}
\examples{
	
data(nir_milk)
data(NIRmilk)

}
\concept{dataset}
\keyword{datasets}
