% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_doby.R
\docType{data}
\name{data_shoes}
\alias{data_shoes}
\alias{shoes}
\alias{shoes_long}
\title{Wear of shoes}
\format{
This data frame contains:
\describe{
\item{A:}{Wear, material A}
\item{B:}{Wear, material B}
\item{boy:}{Id of boy}
\item{footA:}{The foot with material A}
}

An object of class \code{data.frame} with 10 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 4 columns.
}
\usage{
shoes

shoes_long
}
\description{
Wear of soles of shoes of materials A and B for one foot each for of ten boys.
}
\details{
The shoes data are measurements of the amount wear of the
soles of shoes worn by 10 boys. The soles were made to two
different synthetic materials, a standard material A and a
cheaper material B.
}
\references{
Box, Hunter, Hunter (2005) Statistics for Experimenters, 2nd edition Wiley, p. 81.
}
\concept{dataset}
\keyword{datasets}
