% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descStat.R
\name{descStat}
\alias{descStat}
\title{Computing simple descriptive statistics of a numeric vector.}
\usage{
descStat(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Should missing values be removed}
}
\value{
A vector with named elements.
}
\description{
Computing simple descriptive statistics of a numeric
    vector - not unlike what proc means of SAS does
}
\examples{

x <- c(1, 2, 3, 4, NA, NaN)
descStat(x)

}
\seealso{
\code{\link{summaryBy}}
}
\author{
Gregor Gorjanc; gregor.gorjanc <at> bf.uni-lj.si
}
\keyword{utilities}
