% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linest_compute.R
\name{.getLb}
\alias{.getLb}
\title{cat(".getLb")
print(attributes(L))
print(off)}
\usage{
.getLb(L, bhat, VV, ddf.vec, is.est, level = 0.95)
}
\description{
cat(".getLb")
print(attributes(L))
print(off)
}
