% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doby_utilities.R
\name{.rhsf2list}
\alias{.rhsf2list}
\title{Convert right hand sided formula to a list}
\usage{
.rhsf2list(f)
}
\arguments{
\item{f}{A right hand sided formula}
}
\description{
Convert right hand sided formula to a list
}
