% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doby.R
\docType{data}
\name{fatacid}
\alias{fatacid}
\title{Fish oil in pig food}
\format{
A dataframe.
}
\usage{
fatacid
}
\description{
Fish oil in pig food
}
\details{
A fish oil fatty acid \code{X14} has been added in
    different concentrations to the food for pigs in a
    study. Interest is in studying how much of the fatty acid can
    be found in the tissue. The concentrations of \code{x14} in the
    food are \code{verb+dose+=\{0.0, 4.4, 6.2, 9.3\}}.

The pigs are fed with this food until their weight is 60 kg.  From
thereof and until they are slaughtered at 100kg, their food does
not contain the fish oil.
At 60kg (sample=1) and 100kg (sample=2) muscle
biopsies are made and the concentration of x14 is
determined.
Measurements on the same pig are correlated, and pigs are additionally
related through litters.
}
\references{
Data courtesy of Charlotte Lauridsen, Department of
    Animal Science, Aarhus University, Denmark.
}
\keyword{datasets}
