% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_ops.R
\name{formula_ops}
\alias{formula_ops}
\alias{formula_add}
\alias{formula_poly}
\alias{formula_nth}
\alias{formula_to_interaction_matrix}
\alias{formula_chr_to_form}
\alias{any2_rhs_frm}
\alias{any2_lhs_frm}
\alias{any2_rhs_chr}
\alias{any2_lhs_chr}
\title{Formula operations and coercion.}
\usage{
formula_add(f1, f2)

formula_poly(x1, n, noint = FALSE)

formula_nth(f1, n)

formula_to_interaction_matrix(f1)

formula_chr_to_form(rhs, lhs = character(0))

any2_rhs_frm(object)

any2_lhs_frm(object)

any2_rhs_chr(object, string = FALSE)

any2_lhs_chr(object, string = FALSE)
}
\arguments{
\item{f1, f2}{Formulas to be coerced to character vectors.}

\item{x1}{Character vector to be coerced to formulas.}

\item{n}{Positive integer.}

\item{noint}{Boolean.}

\item{rhs, lhs}{right-hand-side and left-hand-side for formula (as characters)}

\item{object}{Character vector or formula.}

\item{string}{Boolean.}
}
\description{
Formula operations and coercion as a supplement to \code{update.formula()}
}
\examples{

formula_poly("z", 2)
formula_poly("z", 2, noint=TRUE)

any2_rhs_chr(c("a", "b", "z"))
any2_rhs_chr(c("a*b", "z"))

any2_rhs_chr(y~a+b+z)
any2_rhs_chr(y~a+b+z, string=TRUE)
any2_rhs_chr(y~a+b+z)
any2_rhs_chr(y~a*b+z)
any2_rhs_chr(y~a*b+z, string=TRUE)

any2_lhs_chr(y~a*b+z)
any2_lhs_chr(log(y)~a*b+z)      ## Not what one might expect
any2_lhs_chr(cbind(y, u)~a*b+z) ## Not what one might expect

formula_chr_to_form(c("a*b", "z"))
formula_chr_to_form(c("a*b", "z"), "y")
formula_chr_to_form(c("a*b", "z"), "log(y)")

formula_add(y~a*b+z, ~-1)
formula_add(y~a*b+z, ~a:b)

}
