% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stability.r
\name{get_formulas}
\alias{get_formulas}
\title{Get formulas from model_stability_glm_class object}
\usage{
get_formulas(object, unique = TRUE, text = FALSE)
}
\arguments{
\item{object}{A model_stability_glm_class object}

\item{unique}{If TRUE, return unique models}

\item{text}{If TRUE, return text (rather than formula).}
}
\description{
Get formulas from model_stability_glm_class object
}
\note{
This is a recent addition. Suggestions and comments are
welcome. The addition may be removed without warning.
}
