% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.R
\name{head_matrix}
\alias{head_matrix}
\alias{head2}
\alias{tail2}
\title{head and tail for matrices}
\usage{
head2(x, n = 6, m = n)

tail2(x, n = 6, m = n)
}
\arguments{
\item{x}{matrix}

\item{n, m}{number of rows and columns}
}
\description{
head and tail for matrices
}
\examples{
M <- matrix(1:20, nrow=4)
head2(M)
head2(M, 2)
}
