% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linestGet.R
\name{linest-get}
\alias{get_X}
\alias{get_X.default}
\alias{get_X.merMod}
\alias{get_contrasts}
\alias{get_contrasts.default}
\alias{get_contrasts.merMod}
\alias{get_vartypes}
\alias{get_xlevels}
\alias{get_xlevels.default}
\alias{get_xlevels.mer}
\alias{get_xlevels.merMod}
\alias{linest-get}
\alias{set_covariate_val}
\alias{set_xlevels}
\title{Auxillary functions for computing lsmeans, contrasts etc}
\usage{
get_xlevels(obj)

\method{get_xlevels}{default}(obj)

\method{get_xlevels}{mer}(obj)

\method{get_xlevels}{merMod}(obj)

get_contrasts(obj)

\method{get_contrasts}{default}(obj)

\method{get_contrasts}{merMod}(obj)

set_xlevels(xlev, at)

get_vartypes(obj)

set_covariate_val(xlev, covariateVal)

get_X(obj, newdata, at = NULL)

\method{get_X}{default}(obj, newdata, at = NULL)

\method{get_X}{merMod}(obj, newdata, at = NULL)
}
\arguments{
\item{obj}{An R object}

\item{xlev}{FIXME: to be described}

\item{at}{FIXME: to be described}

\item{covariateVal}{FIXME: to be described}

\item{newdata}{FIXME: to be described}
}
\description{
Auxillary functions for computing lsmeans, contrasts etc.
}

