% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linest_compute.R
\name{linest}
\alias{linest}
\alias{linest.lm}
\alias{linest.glm}
\alias{linest.geeglm}
\alias{linest.lmerMod}
\alias{linest.merMod}
\alias{linest.default}
\title{Compute linear estimates}
\usage{
linest(object, K = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Model object}

\item{K}{Either \code{NULL} or a matrix with p columns where p is
the number of parameters in the systematic effects in the
model. If \code{NULL} then \code{K} is taken to be the p times
p identity matrix}

\item{level}{The level of the (asymptotic) confidence interval.}

\item{...}{Additional arguments; currently not used.}
}
\value{
A dataframe with results from computing the contrasts.
}
\description{
Compute linear estimates for a range of models. One example of
    linear estimates is population means (also known as LSMEANS).
}
\examples{


## Make balanced dataset
dat.bal <- expand.grid(list(AA=factor(1:2), BB=factor(1:3), CC=factor(1:3)))
dat.bal$y <- rnorm(nrow(dat.bal))

## Make unbalanced dataset
#   'BB' is nested within 'CC' so BB=1 is only found when CC=1
#   and BB=2,3 are found in each CC=2,3,4
dat.nst <- dat.bal
dat.nst$CC <-factor(c(1,1,2,2,2,2,1,1,3,3,3,3,1,1,4,4,4,4))

mod.bal  <- lm(y ~ AA + BB * CC, data=dat.bal)
mod.nst  <- lm(y ~ AA + BB : CC, data=dat.nst)

K <- linest_matrix(mod.nst, effect=c("BB", "CC"))
linest( mod.nst, K )

}
\seealso{
\code{\link{LSmeans}}, \code{\link{linest_matrix}}
}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\keyword{utilities}
