% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_op.R
\name{matrix_op}
\alias{matrix_op}
\title{Matrix operations based on matching dimensions}
\usage{
matrix_op(m1, m2, op = `+`)
}
\arguments{
\item{m1, m2}{matrices with dimnames}

\item{op}{the operation to be performed}
}
\description{
Matrix operations based on matching dimensions
}
\examples{
nms1 <- letters[1:4]
mat1 <- matrix(1:8, nrow=4, ncol=4, dimnames=list(nms1, nms1))

nms2 <- letters[2:3]
mat2 <- matrix(11:18, nrow=2, ncol=4, dimnames=list(nms2, nms1))

matrix_op(mat1, mat2)
matrix_op(mat1, mat2, `*`)

}
