% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbenchmark.R
\name{mb_summary}
\alias{mb_summary}
\alias{summary_mb}
\title{Fast summary of microbenchmark object}
\usage{
mb_summary(object, unit, add.unit = TRUE, ...)

summary_mb(object, unit, add.unit = TRUE, ...)
}
\arguments{
\item{object}{A microbenchmark object}

\item{unit}{The time unit to be used}

\item{add.unit}{Should time unit be added as column to resulting
dataframe.}

\item{...}{Additional arguments; currently not used.}
}
\description{
Fast summary of microbenchmark object. The default
summary method from the microbenchmark package is fairly slow
in producing a summary (due to a call to a function from the
multcomp package.)
}
