% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_doby.R
\docType{data}
\name{nir_milk}
\alias{nir_milk}
\title{nir_milk}
\format{
A list with two components
x Datafrane with infra red light amount at different wavelengths (column names are the wavelengths; just remove the leading X).
y Datafrane with response variables fat, protein,
lactose and dm (drymatter)
}
\usage{
nir_milk
}
\description{
Near infra red light (NIR) measurements are made at
152 wavelengths on 17 milk samples.  While milk runs through a
glass tube, infra red light is sent through the tube and the
amount of light passing though the tube is measured at
different wavelengths.  Each milk sample was additionally
analysed for fat, lactose, protein and dry matter.
}
\examples{
	
data(nir_milk)

}
\seealso{
\link{NIRmilk}
}
\keyword{datasets}
