% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.r
\name{plot_lm}
\alias{plot_lm}
\title{Plot linear model object}
\usage{
plot_lm(lm_fit, format = "2x2")
}
\arguments{
\item{lm_fit}{An object of class 'lm'}

\item{format}{The format of the plot (or a list of plots if format is "list")}
}
\description{
Plot linear model object
}
\examples{

m1 <- lm(speed ~ dist, data=cars)
plot_lm(m1)
plot_lm(m1, "2x2")
plot_lm(m1, "1x4")
plot_lm(m1, "4x1")
plot_lm(m1, "list")

}
