% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{pow}
\alias{pow}
\title{Power function}
\usage{
pow(x, p)
}
\arguments{
\item{x}{An object for which x^p makes sense}

\item{p}{A power}
}
\description{
A function returning x raised to the power p.
}
\author{
Søren Højsgaard
}
