% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_formula2.R
\name{predict.sda_formula}
\alias{predict.sda_formula}
\title{Predict from an sda_formula Model}
\usage{
\method{predict}{sda_formula}(object, newdata, verbose = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{sda_formula}, normally returned by
\code{\link[=sda_formula]{sda_formula()}}.}

\item{newdata}{An optional data frame containing observations for which
predictions are required. Factor levels must be compatible with those
used to fit the model.}

\item{verbose}{Logical. Should the underlying \code{sda} prediction method print
progress information? The default is \code{FALSE}.}

\item{...}{Additional arguments passed to the prediction method for \code{sda}
objects.}
}
\value{
The prediction object returned by the prediction method for \code{sda}
objects.
}
\description{
Converts \code{newdata} to the same design-matrix representation used to fit the
model and then invokes the prediction method for \code{sda} objects. If
\code{newdata} is omitted, predictions are obtained for the training data.
}
\examples{
fit <- sda_formula(Species ~ ., data. = iris)
predict(fit, newdata = iris[1:5, ])

}
