% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{reciprocal}
\alias{reciprocal}
\title{Reciprocal function}
\usage{
reciprocal(x)
}
\arguments{
\item{x}{An R object for which 1/x makes sense}
}
\description{
A function returning the reciprocal of its argument
}
\author{
Søren Højsgaard
}
