% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-scaleBy.R
\name{scale2}
\alias{scale2}
\title{Scale a dataframe or matrix}
\usage{
scale2(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{dataframe or matrix}

\item{center}{Logical, should data be centered.}

\item{scale}{Logical, should data be scaled.}
}
\value{
An object of same class as `x`
}
\description{
Similar to 'base::scale' but scales / centers only numeric values
in data.
}
\examples{

scale2(iris)
}
