% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stability.r
\name{set_list_set_matrix}
\alias{set_list_set_matrix}
\alias{set_list2matrix}
\alias{matrix2set_list}
\title{Matrix representatation of list of vectors and vice versa}
\usage{
set_list2matrix(set_list, aggregate = FALSE)

matrix2set_list(set_matrix)
}
\arguments{
\item{set_list}{list of vectors}

\item{aggregate}{should the vectors be aggregated}

\item{set_matrix}{matrix representatation}
}
\description{
Matrix representatation of list of vectors and vice versa
}
\note{
This is a recent addition. Suggestions and comments are
welcome. The addition may be removed without warning.
}
\examples{
l <- list(c(1,2,3), c(3,2,4), c(3,2,4))
m1 <- set_list2matrix(l)
m1
matrix2set_list(m1)

m2 <- set_list2matrix(l, aggregate=TRUE)
m2
matrix2set_list(m2)


}
