% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy-esticon}
\alias{tidy-esticon}
\alias{tidy.esticon_class}
\title{Tidy an esticon object}
\usage{
\method{tidy}{esticon_class}(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{A 'esticon_class' object (produced by \code{esticon} methods).}

\item{conf.int}{Should confidence intervals be added.}

\item{conf.level}{Desired confidence level.}

\item{...}{Additional arguments; currently not used.}
}
\description{
Tidy summarizes information about the components of the object.
}
