% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy-linest}
\alias{tidy-linest}
\alias{tidy.linest_class}
\title{Tidy a linest object}
\usage{
\method{tidy}{linest_class}(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{A 'linest_class' object (produced by \code{linest} methods).}

\item{conf.int}{Should confidence intervals be added.}

\item{conf.level}{Desired confidence level.}

\item{...}{Additional arguments; currently not used.}
}
\description{
Tidy summarizes information about the components of the object.
}
