% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doby_utilities.R
\name{truncate0}
\alias{truncate0}
\title{Truncate values in a matrix / vector to zero if they are below a certain threshold.}
\usage{
truncate0(x, tol = 0.6, sparse = TRUE)
}
\arguments{
\item{x}{matrix / vector}

\item{tol}{threshold}

\item{sparse}{logical; if TRUE and \code{x} is a matrix, return a sparse matrix}
}
\description{
Truncate values in a matrix / vector to zero if they are below a certain threshold.
}
