% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtools.R
\name{v}
\alias{v}
\title{Shorthand for vparse()}
\usage{
v(...)
}
\arguments{
\item{...}{Variable input in any accepted \code{vparse()} form}
}
\value{
A character vector of variable names
}
\description{
A short and convenient alias for \code{vparse()}. Accepts unquoted names, character vectors, or a formula.
}
