% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtools.R
\name{vcheck}
\alias{vcheck}
\title{Check if variables exist in a data frame}
\usage{
vcheck(df, ...)
}
\arguments{
\item{df}{A data frame}

\item{...}{Variables to check}
}
\value{
TRUE if all variables exist, otherwise error
}
\description{
Check if variables exist in a data frame
}
