% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtools.R
\name{vmap}
\alias{vmap}
\title{Apply a function to each parsed variable name}
\usage{
vmap(.vars, .f)
}
\arguments{
\item{.vars}{Variables to parse}

\item{.f}{Function to apply to each name}
}
\value{
A list of results
}
\description{
Apply a function to each parsed variable name
}
