% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtools.R
\name{vrename}
\alias{vrename}
\title{Rename columns in a data frame}
\usage{
vrename(df, rename_map)
}
\arguments{
\item{df}{A data frame}

\item{rename_map}{A named character vector (old_name = new_name)}
}
\value{
A data frame with renamed columns
}
\description{
Rename columns in a data frame
}
