% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtools.R
\name{vselect}
\alias{vselect}
\title{Select columns from a data frame using flexible input}
\usage{
vselect(df, ...)
}
\arguments{
\item{df}{A data frame}

\item{...}{Variable input (unquoted, character vector, or formula)}
}
\value{
A data frame with selected columns
}
\description{
Select columns from a data frame using flexible input
}
