% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BY-summaryBy2.R
\name{by-summary}
\alias{by-summary}
\alias{summaryBy}
\alias{summary_by}
\title{Function to calculate groupwise summary statistics}
\usage{
summary_by(data, formula, FUN)

summaryBy(formula, data, FUN)
}
\arguments{
\item{data}{A data frame}

\item{formula}{A formula object, see examples below}

\item{FUN}{A list of functions to be applied, see examples below.}
}
\value{
A data frame
}
\description{
Function to calculate groupwise summary statistics, much like
    the summary procedure of SAS
}
\details{
Extra arguments ('...') are passed onto the functions in FUN. Hence care
must be taken that all functions in FUN accept these arguments - OR one can
explicitly write a functions which get around this.  This can particularly
be an issue in connection with handling NAs. See examples below.
Some code for this function has been suggested by Jim Robison-Cox. Thanks.
}
\examples{

data(dietox)
dietox12    <- subset(dietox,Time==12)

fun <- function(x){
  c(m=mean(x), v=var(x), n=length(x))
}

summaryBy(cbind(Weight, Feed) ~ Evit + Cu, data=dietox12,
          FUN=fun)

summaryBy(list(c("Weight","Feed"), c("Evit","Cu")), data=dietox12,
          FUN=fun)

summaryBy(cbind(Weight, Feed) ~ Evit + Cu, data=subset(dietox, Time > 1),
   FUN=fun)

## Calculations on transformed data:

summaryBy(cbind( log(Weight), Feed) ~ Evit + Cu, data=dietox12, FUN=mean)
#' 
## There are missing values in the 'airquality' data, so we remove these
## before calculating mean and variance with 'na.rm=TRUE'. However the
## length function does not accept any such argument. Hence we get
## around this by defining our own summary function in which length is
## not supplied with this argument while mean and var are:

sumfun <- function(x, ...){
  c(m=mean(x, na.rm=TRUE, ...), v=var(x, na.rm=TRUE, ...), l=length(x))
}
summaryBy(Ozone + Solar.R ~ Month, data=airquality, FUN=sumfun )

## Using '.' on the right hand side of a formula means to stratify by
## all variables not used elsewhere:

data(warpbreaks)
summaryBy(breaks ~ wool + tension, warpbreaks, FUN=mean)
summaryBy(breaks ~ ., warpbreaks, FUN=mean)
summaryBy(. ~ wool + tension, warpbreaks, FUN=mean)

}
\author{
Søren Højsgaard, \email{sorenh@math.aau.dk}
}
\seealso{
\code{\link{ave}}, \code{\link{descStat}}, \code{\link{orderBy}},
    \code{\link{splitBy}}, \code{\link{transformBy}}
}
\keyword{univar}

