% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA-doby.R
\docType{data}
\name{crime_rate}
\alias{crime_rate}
\title{crime_rate}
\format{
This data frame contains:
\describe{
\item{Murder:}{crime of murder}
\item{Rape:}{}
\item{Robbery:}{}
\item{Assault:}{}
\item{Burglary:}{residential theft}
\item{Larceny:}{unlawful taking of personal property (pocket picking)}
\item{AutoTheft:}{}
}
}
\usage{
crime_rate
}
\description{
Crime rates per 100,000 inhabitants in states of the USA for different crime
types in 1977.
}
\examples{
	
data(crime_rate)

}
\keyword{datasets}
