% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doby.R
\docType{data}
\name{crimeRate}
\alias{crimeRate}
\title{crimeRate}
\format{
This data frame contains:
\describe{
\item{State:}{State of the USA}
\item{Murder:}{crime of murder}
\item{Rape:}{}
\item{Robbery:}{}
\item{Assault:}{}
\item{Burglary:}{residential theft}
\item{Larceny:}{unlawful taking of personal property (pocket picking)}
\item{AutoTheft:}{}
}
}
\usage{
crimeRate
}
\description{
Crime rates per 100,000 inhabitants in states of the USA for different crime
types.
}
\examples{
	
data(crimeRate)

}
\keyword{datasets}
