% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stability.r
\name{cv_glm_fitlist}
\alias{cv_glm_fitlist}
\title{Cross-validation for list of glm objects}
\usage{
cv_glm_fitlist(data., fit_list, K = 10)
}
\arguments{
\item{data.}{A data frame}

\item{fit_list}{A list of glm objects}

\item{K}{Number of folds}
}
\description{
Cross-validation for list of glm objects
}
