\name{internal}
\alias{internal}

\alias{asNumericMatrix2}
\alias{matrix2dataFrame2}
\alias{subsAttr2}
\alias{doby.xtabs}

\alias{popMatrix}
\alias{linMatrix}

\alias{recodevar}

\alias{print.conMatrix}
\alias{summary.conMatrix}

\alias{linMeans}
\alias{linMeans.lm}

\alias{funBy}

\alias{HTMLreport}
\alias{Rscript2HTML}
\alias{RweaveHTMLreport}
\alias{RweaveHTMLreportSetup}
\alias{RweaveHTMLreportWritedoc}
\alias{RweaveHTMLreportFinish}



\title{Internal functions for the doBy package}
\description{Internal functions for the doBy package}
\keyword{internal}% at least one, from doc/KEYWORDS
