\name{internal}
\alias{internal}

\alias{LSmatrix-class}

\alias{asNumericMatrix2}
\alias{matrix2dataFrame2}
\alias{subsAttr2}
\alias{doby.xtabs}

\alias{print.linearEstimate}

\alias{LSmatrix}
\alias{LSmatrix.default}
\alias{print.LSmatrix}

\alias{funBy}

\alias{print.lmBy}
\alias{print.summary_lmBy}

\alias{specialize}

\alias{print.splitByData}
\alias{get_contrasts}
\alias{get_contrasts.default}
\alias{get_contrasts.merMod}
\alias{get_vartypes}
\alias{get_X}
\alias{get_X.merMod}
\alias{get_X.default}
\alias{get_xlevels}
\alias{get_xlevels.default}
\alias{get_xlevels.mer}
\alias{get_xlevels.merMod}
\alias{set_covariate_val}
\alias{set_xlevels}
%% \alias{}
%% \alias{}
%% \alias{}
%% \alias{}

\alias{get_contrasts.default}
\alias{get_contrasts.merMod}
\alias{get_X.default}
\alias{get_xlevels.default}
\alias{get_xlevels.mer}
\alias{get_xlevels.merMod}
\alias{get_X.merMod}


  
     




\title{Internal functions for the doBy package}
\description{Internal functions for the doBy package}
\keyword{internal}% at least one, from doc/KEYWORDS



%% \alias{linestMatrix}
%% \alias{linestMatrix.default}

%% \alias{print.linestMatrix}
%% \alias{print.linest}

%%\alias{pdiff}

%\alias{}
%% \alias{nullBasis}
%% \alias{nullBasis.lm}
%% \alias{nullBasis.glm}
%% \alias{nullBasis.geeglm}
%% \alias{nullBasis.lmerMod}
%% \alias{nullBasis.matrix}
%% \alias{nullBasis.Matrix}

%% \alias{null_basis}
%%\alias{is_estimable}
%% \alias{get_null_basis.matrix}
%% \alias{get_null_basis.Matrix}
%% \alias{get_null_basis.default}

% \alias{HTMLCommand2}
% \alias{HTMLdefault2}
% \alias{RweaveHTML2}
% \alias{RweaveHTMLRuncode2}

% \alias{RweaveHTMLreport}
% \alias{RweaveHTMLreportSetup}
% \alias{RweaveHTMLreportWritedoc}
% \alias{RweaveHTMLreportFinish}

