\name{internal}
\alias{internal}

\alias{asNumericMatrix2}
\alias{matrix2dataFrame2}
\alias{subsAttr2}
\alias{doby.xtabs}

\alias{popMatrix}
\alias{linMatrix}

\alias{print.conMatrix}
\alias{summary.conMatrix}

\alias{linMeans}
\alias{linMeans.default}
\alias{linMeans.lme}

\alias{funBy}

\alias{print.lmBy}
\alias{print.summary.lmBy}

% \alias{HTMLCommand2}
% \alias{HTMLdefault2}
% \alias{RweaveHTML2}
% \alias{RweaveHTMLRuncode2}

% \alias{RweaveHTMLreport}
% \alias{RweaveHTMLreportSetup}
% \alias{RweaveHTMLreportWritedoc}
% \alias{RweaveHTMLreportFinish}


\alias{specialize}

% deprecated functions
% \alias{recodevar}
% \alias{HTMLreport}
% \alias{Rscript2HTML}

%\alias{}
\alias{print.splitByData}





\title{Internal functions for the doBy package}
\description{Internal functions for the doBy package}
\keyword{internal}% at least one, from doc/KEYWORDS
