\name{internal}
\alias{internal}

\alias{asNumericMatrix2}
\alias{matrix2dataFrame2}
\alias{subsAttr2}
\alias{doby.xtabs}

\alias{popMatrix}
\alias{linMatrix}

\alias{print.conMatrix}
\alias{summary.conMatrix}

\alias{linMeans}
\alias{linMeans.lm}


\title{Internal functions for the doBy package}
\description{Internal functions for the doBy package}
\keyword{internal}% at least one, from doc/KEYWORDS
