\name{internal}
\alias{internal}

%\alias{}
\alias{nullBasis}
\alias{nullBasis.lm}
\alias{nullBasis.glm}
\alias{nullBasis.geeglm}
\alias{nullBasis.lmerMod}
\alias{nullBasis.matrix}
\alias{nullBasis.Matrix}

\alias{LSmatrix-class}

\alias{asNumericMatrix2}
\alias{matrix2dataFrame2}
\alias{subsAttr2}
\alias{doby.xtabs}

\alias{print.LSmeans}

\alias{LSmatrix}
\alias{LSmatrix.default}
\alias{print.LSmatrix}

\alias{funBy}

\alias{print.lmBy}
\alias{print.summary.lmBy}

\alias{specialize}

\alias{print.splitByData}

%% \alias{linestMatrix}
%% \alias{linestMatrix.default}

%% \alias{print.linestMatrix}
%% \alias{print.linest}

\alias{pdiff}


\title{Internal functions for the doBy package}
\description{Internal functions for the doBy package}
\keyword{internal}% at least one, from doc/KEYWORDS

% \alias{HTMLCommand2}
% \alias{HTMLdefault2}
% \alias{RweaveHTML2}
% \alias{RweaveHTMLRuncode2}

% \alias{RweaveHTMLreport}
% \alias{RweaveHTMLreportSetup}
% \alias{RweaveHTMLreportWritedoc}
% \alias{RweaveHTMLreportFinish}

