\name{internal}
\alias{internal}

\alias{orderByOLD}
\alias{splitByOLD}
\alias{summaryByOLD}

\alias{linest_matrix_class-class}
\alias{print.linest_class}
\alias{print.esticon_class}
\alias{print.linest_matrix_class}
\alias{summary.linest_matrix_class}

\alias{asNumericMatrix2}
\alias{matrix2dataFrame2}
\alias{subsAttr2}
\alias{doby.xtabs}


\alias{print.lmBy}
\alias{print.summary_lmBy}
\alias{print.splitByData}

\alias{linest_class-class}


\alias{\%>\%}
 
\title{Internal functions for the doBy package}
\description{Internal functions for the doBy package}
\keyword{internal}% at least one, from doc/KEYWORDS





%%\alias{funBy}

%%\alias{pdiff}

%\alias{}
%% \alias{nullBasis}
%% \alias{nullBasis.lm}
%% \alias{nullBasis.glm}
%% \alias{nullBasis.geeglm}
%% \alias{nullBasis.lmerMod}
%% \alias{nullBasis.matrix}
%% \alias{nullBasis.Matrix}

%% \alias{null_basis}
%%\alias{is_estimable}
%% \alias{get_null_basis.matrix}
%% \alias{get_null_basis.Matrix}
%% \alias{get_null_basis.default}

% \alias{HTMLCommand2}
% \alias{HTMLdefault2}
% \alias{RweaveHTML2}
% \alias{RweaveHTMLRuncode2}

% \alias{RweaveHTMLreport}
% \alias{RweaveHTMLreportSetup}
% \alias{RweaveHTMLreportWritedoc}
% \alias{RweaveHTMLreportFinish}

