\name{internal}
\alias{internal}
\alias{changeNames}

\alias{asNumericMatrix2}
%\alias{returnByValues}


% \alias{print.geeglm}
% \alias{plot.geeglm}
% \alias{summary.geeglm}
% \alias{print.summary.geeglm}
% \alias{print.anova.geeglm}
% \alias{residuals.geeglm}
% \alias{model.matrix.geeglm}

% \alias{anovageePrim}
% \alias{anovaPgee}
\alias{esticonCore}


\title{Internal functions for the doBy package}
\description{Internal functions for the doBy package}
\keyword{internal}% at least one, from doc/KEYWORDS
