% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_to_fun.R
\name{expr_to_fun}
\alias{expr_to_fun}
\title{Convert expression into function object.}
\usage{
expr_to_fun(e)
}
\arguments{
\item{e}{expression}
}
\description{
Convert expression into function object.
}
\examples{

ee <- expression(b1 + (b0 - b1)*exp(-k*x) + b2*x)
ff <- expr_to_fun(ee)
}
