% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stability.r
\name{generate_data_list}
\alias{generate_data_list}
\title{Generate data list}
\usage{
generate_data_list(data., K, method = c("subgroups", "resample"))
}
\arguments{
\item{data.}{A data frame}

\item{K}{Number of folds}

\item{method}{Method for generating data}
}
\description{
Generate data list
}
