% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_doby.R
\docType{data}
\name{math_teachers}
\alias{math_teachers}
\title{Height of math teachers}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 2 columns.
}
\usage{
math_teachers
}
\description{
Height of a sample of math teachers in Danish high
schools collected at a continued education day at Mariager Fjord Gymnasium in 2019.

Format:
height: Height in centimeters
sex: male or female
}
\examples{

aggregate(height ~  sex, data=math_teachers, FUN=mean)
aggregate(height ~  sex, data=math_teachers, FUN=function(x) {c(mean=mean(x), sd=sd(x))})

}
\keyword{datasets}
