% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_stability.r
\name{model_stability_glm}
\alias{model_stability_glm}
\title{Model stability for glm objects}
\usage{
model_stability_glm(
  data.,
  model,
  n.searches = 10,
  method = c("subgroups", "resample"),
  ...
)
}
\arguments{
\item{data.}{A data frame}

\item{model}{A glm object}

\item{n.searches}{Number of searches}

\item{method}{Method for generating data}

\item{...}{Additional arguments to be passed to \code{\link{step}}}
}
\description{
Model stability for glm objects
}
