% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA-doby.R
\docType{data}
\name{NIRmilk}
\alias{NIRmilk}
\title{NIRmilk}
\format{
This data frame contains 18 rows and 158 columns.  The first column
is the sample number.  The columns \code{Xwww} contains the infra red light
amount at wavelength \code{www}.  The response variables are fat, protein,
lactose and dm (dry matter).
}
\usage{
NIRmilk
}
\description{
Near infra red light (NIR) measurements are made at 152 wavelengths on 17
milk samples.  While milk runs through a glass tube, infra red light is sent
through the tube and the amount of light passing though the tube is measured
at different wavelengths.  Each milk sample was additionally analysed for
fat, lactose, protein and dry matter.
}
\examples{
	
data(NIRmilk)

}
\keyword{datasets}
