% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_modelling.R
\name{response}
\alias{response}
\title{Get response variable from model}
\usage{
response(object)
}
\arguments{
\item{object}{lm or glm object}
}
\description{
Get response variable from model
}
\examples{
data(cars)
lm1 <- lm(dist ~ speed + I(speed^2), data=cars)
lm1 |> response() |> head()
cars <- cars |> add_pred(lm1)
cars |> head()
cars <- cars |> add_resid(lm1)
cars 
}
