% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_scaleBy.R
\name{scale_df}
\alias{scale_df}
\alias{scale2}
\title{Scaling numerical values}
\usage{
scale_df(x, center = TRUE, scale = TRUE)

scale2(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{dataframe or matrix}

\item{center}{Logical, should data be centered.}

\item{scale}{Logical, should data be scaled.}
}
\value{
An object of same class as \code{x}
}
\description{
Similar to 'base::scale' but applies to scales / centers only numeric values
in data.
}
\details{
\itemize{
\item If \code{x} is not a dataframe, then base::scale is invoked on
\code{x}.
\item Suppose \code{x} is a dataframe. Then base::scale is invoked
on all columns that are numeric, integer or logical.
}
}
\examples{

scale2(iris)
}
