% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_byrow_bycol.R
\name{split_byrow_bycol}
\alias{split_byrow_bycol}
\alias{split_bycol}
\alias{split_byrow}
\title{Split matrix or dataframe into list}
\usage{
split_bycol(x)

split_byrow(x)
}
\arguments{
\item{x}{Matrix or dataframe.}
}
\description{
Split matrix or dataframe into list by columns or by rows
}
